Macro uucore_procs::help_section

help_section!() { /* proc-macro */ }

Reads a section from a file of the util as a str literal.

It reads from the file specified as the second argument, relative to the crate root. The contents of this file are read verbatim, without parsing or escaping. The name of the help file should match the name of the util. I.e. numfmt should have a file called numfmt.md. By convention, the file should start with a top-level section with the name of the util. The other sections must start with 2 # characters. Capitalization of the sections does not matter. Leading and trailing whitespace of each section will be removed.

Example:

# About
Convert numbers from/to human-readable strings

# Long help
This text will be the long help

help_section!("about", "numfmt.md");